Revert "simplify desing by moving logging common code to utility"
This reverts commit 841b44ddbb700be537e1ebdd170cef7ef952fe96.
